requirements.txt
Django>=4.0
djangorestframework>=3.14
python-dotenv>=0.21.0
dj-database-url>=0.5.0